Rename Component Usage report
Decided to rename this report to "Components" and remove the version
count column since it was little more than clutter. Also added a click
handler that links to a search for projects that use the component.
