+++ src/js/views/Main.jsx
@@ -11,7 +11,7 @@ import { NewEntry, OperationsLog } from './OperationsLog/'
 import { Project } from './Project/'
 import { Projects } from './Projects/'
 import {
-  ComponentUsage,
+  Components,
   NamespaceKPIs,
   OutdatedProjects,
   ProjectTypeDefinitions,
@@ -21,7 +21,6 @@ import { UserProfile, UserSettings } from './User'
 
 import { useSettings } from '../settings'
 import { User as UserSchema } from '../schema'
-import { Components } from './Admin/Components'
 
 function Main({ user }) {
   const [globalState] = useContext(Context)
@@ -70,10 +69,7 @@ function Main({ user }) {
                 />
                 <Route path="/ui/projects" element={<Projects user={user} />} />
                 <Route path="/ui/reports" element={<Reports />} />
-                <Route
-                  path="/ui/reports/component-usage"
-                  element={<ComponentUsage />}
-                />
+                <Route path="/ui/reports/components" element={<Components />} />
                 <Route
                   path="/ui/reports/namespace-kpis"
                   element={<NamespaceKPIs />}
