+++ src/js/views/Reports/Reports.jsx
@@ -1,11 +1,11 @@
-import React, { Fragment, useContext, useEffect } from 'react'
-import { Link, Route } from 'react-router-dom'
+import React, { useContext, useEffect } from 'react'
+import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 
 import { Card } from '../../components'
 import { Context } from '../../state'
 
-export { ComponentUsage } from './ComponentUsage'
+export { Components } from './Components'
 export { NamespaceKPIs } from './NamespaceKPIs'
 export { OutdatedProjects } from './OutdatedProjects'
 export { ProjectTypeDefinitions } from './ProjectTypeDefinitions'
@@ -28,8 +28,8 @@ function Reports() {
         <h1 className="text-gray-700 text-lg">{t('reports.available')}</h1>
         <ul className="list-disc list-inside ml-4 text-gray-600">
           <li>
-            <Link to="/ui/reports/component-usage">
-              {t('reports.componentUsage.title')}
+            <Link to="/ui/reports/components">
+              {t('reports.components.title')}
             </Link>
           </li>
           <li>
